requirements.txt
pandas>=2.2,<3.0
numpy>=1.26,<3.0
openpyxl>=3.1,<4.0
geopandas>=1.0,<2.0
shapely>=2.0,<3.0
pyarrow>=15.0,<20.0
python-dotenv>=1.0,<2.0
fiona>=1.9,<2.0
pyproj>=3.6,<4.0
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
